NIFI-4433 Create a separate NAR. Rename the processor to ConsumeKinesis3Stream
