+++ nifi-extension-bundles/nifi-aws-bundle/nifi-aws-abstract-processors/pom.xml
@@ -98,6 +98,11 @@
 
         <dependency>
             <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-aws-regions</artifactId>
+            <version>2.5.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-aws-service-api</artifactId>
         </dependency>
         <dependency>
