+++ nifi-extension-bundles/nifi-aws-bundle/nifi-aws-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -26,14 +26,13 @@ org.apache.nifi.processors.aws.sqs.GetSQS
 org.apache.nifi.processors.aws.sqs.PutSQS
 org.apache.nifi.processors.aws.sqs.DeleteSQS
 org.apache.nifi.processors.aws.lambda.PutLambda
+org.apache.nifi.processors.aws.kinesis.firehose.PutKinesisFirehose
 org.apache.nifi.processors.aws.dynamodb.GetDynamoDB
 org.apache.nifi.processors.aws.dynamodb.PutDynamoDB
 org.apache.nifi.processors.aws.dynamodb.PutDynamoDBRecord
 org.apache.nifi.processors.aws.dynamodb.DeleteDynamoDB
-org.apache.nifi.processors.aws.kinesis.firehose.PutKinesisFirehose
 org.apache.nifi.processors.aws.kinesis.stream.PutKinesisStream
 org.apache.nifi.processors.aws.kinesis.stream.ConsumeKinesisStream
-org.apache.nifi.processors.aws.kinesis.stream.consumev2.ConsumeKinesisStreamV2
 org.apache.nifi.processors.aws.cloudwatch.PutCloudWatchMetric
 org.apache.nifi.processors.aws.ml.translate.StartAwsTranslateJob
 org.apache.nifi.processors.aws.ml.translate.GetAwsTranslateJobStatus
